update: added tab logic woth gatalogo and the voter section
